LMC v10 3/requirements.txt
flask
Flask-SQLAlchemy
flask-login
cryptography
